Read and insert data to grid View
So I've tried to read all the data and split it up so i can then read
the data in grid view. Unfortunately the program builds with no errors
but nothing is output. As I've gone back to look through the code I'm
struggling to figure out the problem, but i have realised that my
variable names are muddled up so going to work on that.